refactor(mespapiers): Make SearchResult dumber and deal only with result
