+++ packages/cozy-mespapiers-lib/src/components/Home/ResultForSearch.jsx
@@ -2,8 +2,13 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
+import Empty from 'cozy-ui/transpiled/react/Empty'
 
-import { buildFilesWithContacts } from '../Papers/helpers'
+import HomeCloud from '../../assets/icons/HomeCloud.svg'
+import {
+  buildFilesWithContacts,
+  getContactsRefIdsByFiles
+} from '../Papers/helpers'
 import { useScannerI18n } from '../Hooks/useScannerI18n'
 import SearchResult from '../SearchResult/SearchResult'
 import { filterPapersByThemeAndSearchValue } from './helpers'
@@ -25,7 +30,30 @@ const ResultForSearch = ({ contacts, papers, selectedTheme, searchValue }) => {
     scannerT
   }).map(({ file }) => file)
 
-  return <SearchResult filteredPapers={filteredPapers} contacts={contacts} />
+  const contactIds = getContactsRefIdsByFiles(filteredPapers)
+  const contactsByIds = contacts.filter(contact =>
+    contactIds.includes(contact._id)
+  )
+
+  const result = buildFilesWithContacts({
+    files: filteredPapers,
+    contacts: contactsByIds,
+    t
+  })
+
+  if (result.length === 0) {
+    return (
+      <Empty
+        className="u-ph-1"
+        icon={HomeCloud}
+        iconSize="large"
+        title={t('Search.empty.title')}
+        text={t('Search.empty.text')}
+      />
+    )
+  }
+
+  return <SearchResult result={result} />
 }
 
 ResultForSearch.propTypes = {
